+++ plugin-maven/src/test/java/com/diffplug/spotless/maven/MavenIntegrationHarness.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016-2023 DiffPlug
+ * Copyright 2016-2024 DiffPlug
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -177,6 +177,10 @@ public class MavenIntegrationHarness extends ResourceHarness {
 
 	protected void writePomWithJsonSteps(String... steps) throws IOException {
 		writePom(groupWithSteps("json", including("**/*.json"), steps));
+	}
+
+	protected void writePomWithShellSteps(String... steps) throws IOException {
+		writePom(groupWithSteps("shell", including("**/*.sh"), steps));
 	}
 
 	protected void writePomWithYamlSteps(String... steps) throws IOException {