+++ plugin-maven/CHANGES.md
@@ -3,6 +3,7 @@
 We adhere to the [keepachangelog](https://keepachangelog.com/en/1.0.0/) format (starting after version `1.27.0`).
 
 ## [Unreleased]
+* Support for `gofmt` ([#2001](https://github.com/diffplug/spotless/issues/2001))
 
 ## [2.42.0] - 2024-01-15
 ### Added