+++ plugin-maven/CHANGES.md
@@ -5,7 +5,7 @@ We adhere to the [keepachangelog](https://keepachangelog.com/en/1.0.0/) format (
 ## [Unreleased]
 ### Added
 * Support for formatting shell scripts via [shfmt](https://github.com/mvdan/sh). ([#1998](https://github.com/diffplug/spotless/issues/1998))
-* Support for `gofmt` ([#2001](https://github.com/diffplug/spotless/issues/2001))
+* Support for `gofmt` ([#2001](https://github.com/diffplug/spotless/pull/2001))
 
 ## [2.42.0] - 2024-01-15
 ### Added